api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
